pre_service/requirements.txt
flask==3.0.3
umbral==2.0.0
charm-crypto==0.50
requests==2.31.0